hsi-12594 Cambia tipo column fhir_quantity.value a numeric

+++ back-end/clinic-history/src/main/java/ar/lamansys/sgh/clinichistory/domain/ips/FhirQuantityBo.java
@@ -6,6 +6,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -14,7 +16,7 @@ import lombok.ToString;
 public class FhirQuantityBo {
 
 	private Integer id;
-	private Float value;
+	private BigDecimal value;
 	private String unit;
 
 	static public FhirQuantityBo empty() {
